+++ ChangeLog.rst
@@ -3,6 +3,7 @@ Unreleased
 
 - (nvie/pip-tools#453) Write relative find-links opts to output file
 - Add "--silent" option for the compile command
+- Rename "prequ compile" to "prequ compile-in"
 
 0.180.8
 -------